Fixes multi-inject bug
No longer will calling `Meteor.typeahead.inject()` twice on the same element cause it to stop working. It simply ignores the second initialization.
Also, if inject is called from within a template instance, it will only search that template's DOM nodes for typeahead elements to be injected. Otherwise, it will search the entire DOM structure, as it did before.